Move DataViewsPicker story
