+++ packages/dataviews/src/dataviews-picker/stories/index.story.tsx
@@ -13,11 +13,11 @@ import { Stack } from '@wordpress/ui';
 /**
  * Internal dependencies
  */
-import DataViewsPicker from '../dataviews-picker/index';
-import { LAYOUT_PICKER_GRID, LAYOUT_PICKER_TABLE } from '../constants';
-import filterSortAndPaginate from '../utils/filter-sort-and-paginate';
-import type { ActionButton, View } from '../types';
-import { data, fields, type SpaceObject } from './dataviews.fixtures';
+import DataViewsPicker from '../index';
+import { LAYOUT_PICKER_GRID, LAYOUT_PICKER_TABLE } from '../../constants';
+import filterSortAndPaginate from '../../utils/filter-sort-and-paginate';
+import type { ActionButton, View } from '../../types';
+import { data, fields, type SpaceObject } from './fixtures';
 
 const meta = {
 	title: 'DataViews/DataViewsPicker',
